Fixed problem with input calculation and variable references.
If a table path defined an input mapping and that original input was
also used as a variable reference in the table, then the original
input and the mapped input would show up in the final list of
calculated inputs.  It should not have contained the original input
in that case.