Changed matching logic with regard to missing/null keys. (#12)
Previously, if a INPUT cell was a blank value, it would only match to
blank value input.  For regular input values this was not an issue
since if you didn't supply an input it was defaulted to blank.  However
intermediate values during the staging process were not handled the
same way.  If the input was not in the context, it would not match.
Now it will match since keys not in the context are handled exactly
the same as blank.

(cherry picked from commit b31c2f9874d60504f9eb63cb0d4bf6ae3477f944)