Changed how outputs are calculated.
If outputs are specified, those were only included in the outputs if
either they had a default or if it was possibly output through the
processing of the mappings.  However the staging code sets outputs
without a default to blank, so it doesn't matter whether a default is
specified.  If the outputs are specified, then that is the exact list
of resulting outputs regardless of the inputs.